lib/chan.[ch]: one-shot semaphore finalisation fix.
m0_clink has "one-shot" mode, where it is removed
from the channel once it is signalled. Existing
implementation removed the clink from the channel
in clink_signal(), called by m0_chan_{signal,
broadcast}(). clink_signal() first fianlised the
semaphore and then UP-ed it, waking up waiters.
This happened to work on Linux, but Darwin
semaphores are not that forgiving. Attempts to
fix this "locally" failed because the following
requirements clearly contradict each other:

    - the semaphore has to be finalised after is
      incremented;

    - after the semaphore was incremented, a
      waiter can immediately free the clink.

The solution is to finalise the semaphore
on the waiter side, either after wake-up or
(in case nobody was sleeping) during clink
finalisation.

An additional clink flag M0_CF_HEARD_BANSHEE
is needed to notify waiers that semaphore
finalisation is needed.